websocket: Enable websocket as heuristic dissector for TCP
Add a new option to use websocket as heuristic dissector for TCP.

This is useful when the connection setup of a HTTP connection which got
upgraded to a websocket was not recorded.

If this option is disabled via preference the resulting behaviour should be
the same as prior. It is disabled by default.