wmem: Add assertion macro with WS_DISABLE_ASSERT
wmem has many assertions during dissection, these are assumed to have
a measurable performance impact so remove assertions with
WS_DISABLE_ASSERT, like is done elsewhere.

We don't use ws_assert() to avoid a dependency on wsutil.

g_assert_not_reached() does not have a performance impact and for
that reason should not be disabled.