WOWW: Add protocol
The protocol is a continuation of the WOW protocol occuring between the
world server and the client (as opposed to the login server and the
client).

The first two opcodes are unencrypted and perform setup for the
encryption.
The encryption was setup in the WOW protocol through SRP6.
Using the session key for encryption like this is not part of the SRP6
protocol.
All other opcodes are encrypted using the session key, which will need
to be deduced first.