tcp: switch packet_scope to pinfo->pool
Per mailing list discussion:
https://www.wireshark.org/lists/wireshark-dev/202107/msg00030.html

Long-term we want to get rid of the wmem_*_scope globals in favour of
passing wmem pools around. Step one is to replace all reasonable uses of
wmem_packet_scope() with pinfo->pool which has effectively the same
lifespan. This converts the TCP dissector as a proof of concept. TCP is
a common enough protocol this should stress-test the idea fairly well.