Proposed rewording of filter tooltips
For #16186. Proposed changes to the tooltips which appear when a filter
expression is potentially problematic.

Rename references to "User's Guide" to "Help" since the link to the
User's Guide in the Help menu is just called *Contents*.

Name specific sections within the help which pertain to the warning
tooltip being shown. Gives first-time users some help in finding the
right part of the sizeable User's Guide.