WOWW: Make undecryptable server messages stop decryption
Since it is (more or less) impossible to know exactly how many messages
are in a PDU we don't know how many times to increment the session key
index, meaning we might decrypt incorrectly.
To prevent this we freeze the decryption of server messages when we meet
an undecryptable message.