btatt: fix reported dissector bug.
For better or worse, currently, if you have a field that is broken into
bitfields, the top-level field must be integral, so the entire field's
value is shown.  A case could be made that FT_NONE should be supported,
but that's not the case now.

Fixes issue #17505.


(cherry picked from commit 436a9fabcb7124cd26c9a7c727fbf384ed5dc1b8)