[#17519] capinfos: no section headers in table output
Remove "Section N:" headers from capinfos table (`-T`) output when
outputting "additional capture file information" with `-F` and/or
comments with `-k`. These headers broke the formatting of table output.

The downside to this fix is that pcapng files with multiple SHBs and/or
comments will have extra table columns that don't line up with the
header, as in:

    ...,hardware1,os1,application1,comment1,comment1,hardware2,os2,
         application2,comment2,comment2,...

There's no real good way around this though.