etw: use PACK_FLAGS_DIRECTION() to extract the direction from the flags.
Use PACK_FLAGS_DIRECTION() rather than making with
PACK_FLAGS_DIRECTION_MASK.  They happen to be at the bottom of the flags
option, so no shifing is needed, but it's cleaner to use the macro.

Make the packet flags variable local to the code that uses it, and fix
indentation (don't use tabs - they're not used elsewhere), while we're
at it.