gsm_a_rr: add NRI decoding when pool is used
Protocol BSSGP has option in GUI to enable decoding NRI when SGSN pool
is used. But this configuration option is not used in code.

I have marked this option as obsolete and add new single option to protocol dissector gsm_a_rr
where TLLI is decoded and added code for processing NRI.