gsm_sim: Fix MANAGE CHANNEL
Use correct offset for P1 and P2 when showing channel operation
and channel number.

According to TS 102 221 V14.2.0 the Open channel operation is using
P2 for channel, and let UICC assign when this is 0. Show P3 as Le.
Make P3 optional because it's not present in at least Close channel
operation.

Change text in Info column to avoid "Channel Channel: x".