lapd: clean up variable names, remove unnecessary initialization.
Use "flags" for the OPT_PKT_FLAGS option value, as we do elsewhere.  Use
"lapd_flags" for the variable holding the LAPD flags.

We don't need to initialize flags, as we extract the OPT_PKT_FLAGS
option into it immediately after initializing it (if the attempt to get
that option fails, it's not set, but it's not used, either).