DVB-S2: add to "decode as" for UDP
The heuristic dissector doesn't always work. It is convenient to have
DVB-S2 in the "decode as" menu as well.

The heuristic dissector does not dissect packets where the modeadapt
cannot be detected (CRC errors or other problems). The "decode as"
dissector, on the other hand, will attempt decoding anyway, using the
preferred mode adaptation set in preferences.