GSM A-bis/OML: Fix false positive on expert notification for ABIS_OM_MDISC_MANUF
As per TS 12.21 section 8.1.4 "Manufacturer-Defined O&M messages",
NOTE 1:
"""
The Length Indicator gives the length of the Manufacturer-defined O&M data field in the message
segment being transported which is less than or equal to 255 octet.
"""

Where the "Manufacturer-defined O&M data field" is the content AFTER the
"ManId Length Indicator" + "Manuf. Identifier" as can be seen in the
table of the same section.

This fix was tested against osmo-bts, which implements the ipaccess
manufacturer extensions.