BT-DHT: Test packets even if the dissector is set
BitTorrent clients use the same UDP conversation for both DHT and
uTP, switching back and forth between the two at connection start.
So even if the dissector has been set for the conversation or
ports to BT-DHT, test the packet and reject it if not DHT in order
to give the uTP dissector a chance. Fix #17626


(cherry picked from commit 5c185238a49bb6b2120d5a32f75ba99e0891919f)