BT-DHT, BT-uTP: Use conversation_set_dissector_from_frame_number
Since the UDP connection switches back and forth between DHT and uTP,
use conversation_set_dissector_from_frame_number so that the dissector
called by try_conversation_dissector in packet-udp.c doesn't change for
a given frame based on the last packet clicked in the GUI.

Split out a heuristic dissector from uTP so that conversation_set_dissector
is only called from the heuristic dissector.

This doesn't make a difference when the heuristics are accurate but
might in some edge cases.